Feat : 내부 api 및 productVector 조회
