--- recommendation/src/main/java/store/_0982/recommendation/presentation/AiInternalController.java
@@ -1,21 +1,0 @@
-package store._0982.recommendation.presentation;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-import store._0982.recommendation.application.PromptService;
-import store._0982.recommendation.presentation.dto.InterestSummaryRequest;
-
-@RestController
-@RequiredArgsConstructor
-public class AiInternalController {
-
-    private final PromptService promptService;
-
-    @PostMapping("/internal/ai/interest-summary")
-    public String summarizeInterest(@RequestBody InterestSummaryRequest request){
-        return promptService.summarizeInterest(request.descriptions());
-    }
-
-}
