+++ recommendation/src/main/java/store/_0982/recommendation/presentation/dto/ProductVectorResponse.java
@@ -1,0 +1,4 @@
+package store._0982.recommendation.presentation.dto;
+
+public record ProductVectorResponse(float[] vector) {
+}
